Rename eunomia-sdk-python to eunomia-sdk
