+++ docs/api/sdks/index.md
@@ -4,7 +4,7 @@ Each SDK is built and distributed as a separate package that can be installed wi
 
 ## Available SDKs
 
-| SDK name   | Package name             | Jump to                                      |
-| ---------- | ------------------------ | -------------------------------------------- |
-| Python     | `eunomia-sdk-python`     | [:material-arrow-right: Page](python.md)     |
-| TypeScript | `eunomia-sdk-typescript` | [:material-arrow-right: Page](typescript.md) |
+| SDK name   | Package name  | Published to | Jump to                                      |
+| ---------- | ------------- | ------------ | -------------------------------------------- |
+| Python     | `eunomia-sdk` | PyPI         | [:material-arrow-right: Page](python.md)     |
+| TypeScript | `eunomia-sdk` | NPM          | [:material-arrow-right: Page](typescript.md) |
