+++ pkgs/extensions/langchain/pyproject.toml
@@ -9,7 +9,7 @@ license = "Apache-2.0"
 readme = "README.md"
 requires-python = ">=3.10"
 dependencies = [
-    "eunomia-sdk-python>=0.3.4",
+    "eunomia-sdk>=0.3.4",
     "langchain>=0.3.20",
     "langchain-core>=0.3.41",
 ]
@@ -22,4 +22,4 @@ build-backend = "hatchling.build"
 packages = ["src/eunomia_langchain"]
 
 [tool.uv.sources]
-eunomia-sdk-python = { workspace = true }
+eunomia-sdk = { workspace = true }
